#32 Move types from chess engine to Types.ts, refactor chess engine
